遗漏人员信息补充

+++ src/main/java/com/yicloud/trans/model/mssql/MzypmxkF.java
@@ -16,7 +16,7 @@ public class MzypmxkF implements Serializable {
     @TableId
     private Long id;
 
-    private Integer zyh;
+    private Long zyh;
 
     private Date rq;
